requirements.txt
aiodns==4.0.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-jinja2==1.6
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-types==0.7.0
anyio==4.13.0
arrow==1.4.0
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
distro==1.9.0
frozenlist==1.8.0
g4f==7.3.8
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-genai==1.70.0
googleapis-common-protos==1.74.0
googlesearch-python==1.3.0
greenlet==3.3.2
groq==1.1.2
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
Jinja2==3.0.3
jinja2-time==0.2.0
loguru==0.7.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
nanoid==2.0.0
nest-asyncio2==1.7.2
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycares==5.0.1
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyjabber==0.4.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2022.1
PyYAML==6.0.3
requests==2.33.1
rich==14.3.3
singletonify==0.2.4
six==1.17.0
slixmpp_multiplatform==1.12.0
sniffio==1.3.1
soupsieve==2.8.3
spade==4.1.2
SQLAlchemy==2.0.48
tenacity==9.1.4
timeago==1.0.16
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
win32_setctime==1.2.0
winloop==0.6.0
yarl==1.23.0
